Revert "Features-dev-002: crear crud para products"
