+++ src/components/pages/inventory/InventoryLogic.jsx
@@ -94,11 +94,7 @@ const InventoryLogic = () => {
   const handleSubmit = (e) => {
     e.preventDefault();
     if (!newItem.nombre || !newItem.stockActual || !newItem.unidad) {
-      Swal.fire({
-        icon: "error",
-        title: "Oops...",
-        text: "Por favor complete los campos requeridos.",
-      });
+      showToast("Error", "Por favor complete los campos requeridos");
       return;
     }
 
@@ -117,7 +113,6 @@ const InventoryLogic = () => {
     handleClose();
   };
 
-  // FUNCION PARA FILTRAR PRODUCTOS
   const handleSearch = (event) => {
     const query = event.target.value;
     filterIngredients(query);
@@ -135,8 +130,8 @@ const InventoryLogic = () => {
     getStockStatus,
     handleEdit,
     handleSearch,
+    search,
     filteredInventory,
-    isEditing,
   };
 
   return (
