+++ src/components/pages/products/Products.jsx
@@ -9,55 +9,145 @@ import {
   Col,
   Modal,
   Badge,
-  InputGroup,
 } from "react-bootstrap";
-import { FaPlus, FaPen, FaTrash, FaSearch } from "react-icons/fa";
+import { FaPlus, FaPen, FaTrash } from "react-icons/fa";
 
-const Products = ({ data }) => {
-  const {
-    showModal,
-    products,
-    inventory,
-    newProduct,
-    removeProduct,
-    handleClose,
-    handleShow,
-    handleInputChange,
-    handleIngredientChange,
-    selectedIngredient,
-    handleSubmit,
-    handleEdit,
-    handleSearch,
-    filteredProduct,
-    getIngredientName,
-    addIngredient,
-    removeIngredient,
-    isEditing,
-  } = data;
+const Products = () => {
+  const [showModal, setShowModal] = useState(false);
+  const [products, setProducts] = useState([]);
+  const [newProduct, setNewProduct] = useState({
+    id: "",
+    nombre: "",
+    precio: "",
+    descripcion: "",
+    proveedor: "",
+    estado: true,
+    ingredientes: [],
+  });
+
+  // Simulación de inventario (esto debería venir del contexto o props)
+  const [inventory] = useState([
+    { id: "pan_001", nombre: "Pan de hamburguesa", unidad: "unidad" },
+    { id: "carne_002", nombre: "Carne de res", unidad: "g" },
+    { id: "queso_003", nombre: "Queso cheddar", unidad: "rebanada" },
+  ]);
+
+  const [selectedIngredient, setSelectedIngredient] = useState({
+    id: "",
+    cantidad: "",
+  });
+
+  const handleClose = () => {
+    setShowModal(false);
+    setNewProduct({
+      id: "",
+      nombre: "",
+      precio: "",
+      descripcion: "",
+      proveedor: "",
+      estado: true,
+      ingredientes: [],
+    });
+    setSelectedIngredient({ id: "", cantidad: "" });
+  };
+
+  const handleShow = () => setShowModal(true);
+
+  const handleInputChange = (e) => {
+    const { name, value } = e.target;
+    setNewProduct((prev) => ({
+      ...prev,
+      [name]: value,
+    }));
+  };
+
+  const handleIngredientChange = (e) => {
+    const { name, value } = e.target;
+    setSelectedIngredient((prev) => ({
+      ...prev,
+      [name]: value,
+    }));
+  };
+
+  const addIngredient = () => {
+    if (!selectedIngredient.id || !selectedIngredient.cantidad) {
+      alert("Por favor seleccione un ingrediente y especifique la cantidad");
+      return;
+    }
+
+    const ingredientFromInventory = inventory.find(
+      (item) => item.id === selectedIngredient.id
+    );
+
+    if (!ingredientFromInventory) return;
+
+    const newIngredient = {
+      id: selectedIngredient.id,
+      cantidad: parseFloat(selectedIngredient.cantidad),
+      unidad: ingredientFromInventory.unidad,
+    };
+
+    // Verificar si el ingrediente ya existe
+    if (newProduct.ingredientes.some((ing) => ing.id === newIngredient.id)) {
+      alert("Este ingrediente ya ha sido agregado");
+      return;
+    }
+
+    setNewProduct((prev) => ({
+      ...prev,
+      ingredientes: [...prev.ingredientes, newIngredient],
+    }));
+
+    setSelectedIngredient({ id: "", cantidad: "" });
+  };
+
+  const removeIngredient = (ingredientId) => {
+    setNewProduct((prev) => ({
+      ...prev,
+      ingredientes: prev.ingredientes.filter((ing) => ing.id !== ingredientId),
+    }));
+  };
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (
+      !newProduct.nombre ||
+      !newProduct.precio ||
+      newProduct.ingredientes.length === 0
+    ) {
+      alert(
+        "Por favor complete todos los campos requeridos y agregue al menos un ingrediente"
+      );
+      return;
+    }
+
+    const productToAdd = {
+      ...newProduct,
+      id: `${newProduct.nombre
+        .toLowerCase()
+        .replace(/\s+/g, "_")}_${Date.now()}`,
+      precio: parseFloat(newProduct.precio),
+      fechaRegistro: new Date().toISOString().split("T")[0],
+      fechaActualizacion: new Date().toISOString().split("T")[0],
+    };
+
+    setProducts((prev) => [...prev, productToAdd]);
+    handleClose();
+  };
+
+  const getIngredientName = (id) => {
+    const ingredient = inventory.find((item) => item.id === id);
+    return ingredient ? ingredient.nombre : id;
+  };
 
   return (
     <>
       <h2 className="mb-4">Productos</h2>
       <Card className="mb-4">
         <Card.Body>
-          <Row className="align-items-center">
-            <Col md={6} className="mb-3 mb-md-0">
-              <Button variant="primary" onClick={handleShow}>
-                <FaPlus className="me-2" /> Agregar Producto
-              </Button>
-            </Col>
-            <Col md={6}>
-              <InputGroup>
-                <InputGroup.Text>
-                  <FaSearch />
-                </InputGroup.Text>
-                <Form.Control
-                  placeholder="Buscar por nombre"
-                  onChange={handleSearch}
-                />
-              </InputGroup>
-            </Col>
-          </Row>
+          <Button variant="primary" onClick={handleShow}>
+            <FaPlus className="me-2" /> Agregar Producto
+          </Button>
         </Card.Body>
       </Card>
 
@@ -77,113 +167,38 @@ const Products = ({ data }) => {
               </tr>
             </thead>
             <tbody>
-              {filteredProduct.length > 0 ? (
-                filteredProduct.map((product, index) => (
-                  <tr key={product.id + index}>
-                    <td>{product.id.slice(0, 4) + "..."}</td>
-                    <td>{product.nombre}</td>
-                    <td>${product.precio}</td>
-                    <td>{product.descripcion.slice(0, 10) + "..."}</td>
-                    <td>{product.proveedor}</td>
-                    <td>
-                      <Badge
-                        bg={
-                          product.estado === "true" || product.estado === true
-                            ? "success"
-                            : "danger"
-                        }
-                      >
-                        {product.estado === "true" || product.estado === true
-                          ? "Activo"
-                          : "Inactivo"}
-                      </Badge>
-                    </td>
-                    <td>
-                      <ul className="list-unstyled m-0">
-                        {product.ingredientes.map((ing) => (
-                          <li key={ing.id}>
-                            {getIngredientName(ing.id)} ({ing.cantidad}{" "}
-                            {ing.unidad})
-                          </li>
-                        ))}
-                      </ul>
-                    </td>
-                    <td>
-                      <Button
-                        variant="warning"
-                        size="sm"
-                        className="me-2"
-                        onClick={() => handleEdit(product)}
-                      >
-                        <FaPen />
-                      </Button>
-                      <Button
-                        variant="danger"
-                        size="sm"
-                        onClick={() => removeProduct(product.id)}
-                      >
-                        <FaTrash />
-                      </Button>
-                    </td>
-                  </tr>
-                ))
-              ) : products.length > 0 ? (
-                products.map((product, index) => (
-                  <tr key={product.id + index}>
-                    <td>{product.id.slice(0, 4) + "..."}</td>
-                    <td>{product.nombre}</td>
-                    <td>${product.precio}</td>
-                    <td>{product.descripcion.slice(0, 10) + "..."}</td>
-                    <td>{product.proveedor}</td>
-                    <td>
-                      <Badge
-                        bg={
-                          product.estado === "true" || product.estado === true
-                            ? "success"
-                            : "danger"
-                        }
-                      >
-                        {product.estado === "true" || product.estado === true
-                          ? "Activo"
-                          : "Inactivo"}
-                      </Badge>
-                    </td>
-                    <td>
-                      <ul className="list-unstyled m-0">
-                        {product.ingredientes.map((ing) => (
-                          <li key={ing.id}>
-                            {getIngredientName(ing.id)} ({ing.cantidad}{" "}
-                            {ing.unidad})
-                          </li>
-                        ))}
-                      </ul>
-                    </td>
-                    <td>
-                      <Button
-                        variant="warning"
-                        size="sm"
-                        className="me-2"
-                        onClick={() => handleEdit(product)}
-                      >
-                        <FaPen />
-                      </Button>
-                      <Button
-                        variant="danger"
-                        size="sm"
-                        onClick={() => removeProduct(product.id)}
-                      >
-                        <FaTrash />
-                      </Button>
-                    </td>
-                  </tr>
-                ))
-              ) : (
-                <tr>
-                  <td colSpan="8" className="text-center">
-                    No hay productos registrados.
+              {products.map((product) => (
+                <tr key={product.id}>
+                  <td>{product.id}</td>
+                  <td>{product.nombre}</td>
+                  <td>${product.precio.toFixed(2)}</td>
+                  <td>{product.descripcion}</td>
+                  <td>{product.proveedor}</td>
+                  <td>
+                    <Badge bg={product.estado ? "success" : "danger"}>
+                      {product.estado ? "Activo" : "Inactivo"}
+                    </Badge>
+                  </td>
+                  <td>
+                    <ul className="list-unstyled m-0">
+                      {product.ingredientes.map((ing) => (
+                        <li key={ing.id}>
+                          {getIngredientName(ing.id)} ({ing.cantidad}{" "}
+                          {ing.unidad})
+                        </li>
+                      ))}
+                    </ul>
+                  </td>
+                  <td>
+                    <Button variant="warning" size="sm" className="me-2">
+                      <FaPen />
+                    </Button>
+                    <Button variant="danger" size="sm">
+                      <FaTrash />
+                    </Button>
                   </td>
                 </tr>
-              )}
+              ))}
             </tbody>
           </Table>
         </Card.Body>
@@ -192,11 +207,7 @@ const Products = ({ data }) => {
       {/* Modal para agregar producto */}
       <Modal show={showModal} onHide={handleClose} size="lg">
         <Modal.Header closeButton>
-          <Modal.Title>
-            {isEditing === true
-              ? "Actualizar Producto"
-              : "Agregar Nuevo Producto"}
-          </Modal.Title>
+          <Modal.Title>Agregar Nuevo Producto</Modal.Title>
         </Modal.Header>
         <Modal.Body>
           <Form onSubmit={handleSubmit}>
@@ -279,8 +290,8 @@ const Products = ({ data }) => {
                         onChange={handleIngredientChange}
                       >
                         <option value="">Seleccionar...</option>
-                        {inventory.map((item, index) => (
-                          <option key={item.id + index} value={item.id}>
+                        {inventory.map((item) => (
+                          <option key={item.id} value={item.id}>
                             {item.nombre} ({item.unidad})
                           </option>
                         ))}
@@ -347,15 +358,9 @@ const Products = ({ data }) => {
               >
                 Cancelar
               </Button>
-              {isEditing === true ? (
-                <Button variant="primary" type="submit">
-                  Actualizar Producto
-                </Button>
-              ) : (
-                <Button variant="primary" type="submit">
-                  Guardar Producto
-                </Button>
-              )}
+              <Button variant="primary" type="submit">
+                Guardar Producto
+              </Button>
             </div>
           </Form>
         </Modal.Body>
