+++ src/context/InventoryContext.jsx
@@ -74,7 +74,7 @@ const InventoryProvider = ({ children }) => {
     const result = await Swal.fire({
       title: "¿Estás seguro?",
       text: "¡No podrás revertir esto!",
-      icon: "warning",
+      icon: "warning", // 🔹 Se corrigió el icono
       showCancelButton: true,
       confirmButtonColor: "#3085d6",
       cancelButtonColor: "#d33",
